Add files via upload
1. A data science student is using two machine learning algorithms to analyze a given
dataset. First, she tests her model with a Decision Tree using a 10-fold Cross Validation
and records the Accuracy for each fold. She then repeats the same process using a 12-fold
Logistic Regression algorithm on the same dataset. The accuracy for each fold of the
algorithms are given as follows:
Decision Tree: 75, 79, 69, 78, 65, 87, 74, 81, 77, 88
Logistic Regression: 85, 68, 78, 73, 69, 76, 69, 80, 73, 67, 78, 82
(a) Is the dataset paired or unpaired?
(b) Can she conclude that either one of the machine learning algorithm is more accurate
than the other?
i. Create vectors for each algorithm. Check the means of the accuracy reported
for the two models.
ii. Run wilcox.test command to check the significance of the comparison.